requirements.txt
# Core dependency for YAML config parsing.
PyYAML>=6.0
